+++ test/bencodelix_test.exs
@@ -34,7 +34,7 @@ defmodule BencodelixTest do
   end
 
   test " decode susccessive string" do
-    assert(Bencode.decode("6:azerty6:azerty") == ["azerty","azerty"])
+    assert(Bencode.decode("6:qwerty6:azerty") == ["qwerty","azerty"])
   end
   test "encode  dic" do
     dict = HashDict.new()
@@ -63,14 +63,17 @@ d2:ex45:class clojure.lang.Compiler$CompilerException7:root-ex45:class clojure.l
   end
 
   test "decode complex datas" do
-    s = """
+    # when miltiple values are concatened we return a list
+    sin = """
 d2:ex45:class clojure.lang.Compiler$CompilerException7:root-ex45:class clojure.lang.Compiler$CompilerException7:session36:3ea0cb6b-47ff-4105-b308-a4f26774cdf86:statusl10:eval-erroreed3:err123:CompilerException java.lang.RuntimeException: Unable to resolve symbol: e in this context, compiling:(NO_SOURCE_PATH:0:0) 
 ...(7)> 7:session36:3ea0cb6b-47ff-4105-b308-a4f26774cdf8ed2:ns4:user7:session36:3ea0cb6b-47ff-4105-b308-a4f26774cdf85:value1:8ed7:session36:3ea0cb6b-47ff-4105-b308-a4f26774cdf86:statusl4:doneee
 """
-    r = Bencode.decode(s)
-    IO.puts ("Complex datas #{inspect r} ")
-    e = Bencode.encode(r)
-    assert(e == s )
+    l = Bencode.decode(sin)
+    # IO.puts ("Complex datas #{inspect l} ")
+    e = Bencode.encode(l)
+    n = Bencode.decode(e)
+    # when re-decode the  encoded decoded should be same
+    assert(l == n )
   end
 end