+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # CHANGELOG for Elicloj
+## V 0.0.2 ##
+- now Regex handle multiline
+- when multiple thing concatened on a input line they are decoded as a list
 ## V 0.0.1 ##
 - initial release